inprogress-seperation of avatar section
